Reorganized | whole package scripts
