+++ argprism/__init__.py
@@ -1,6 +1,6 @@
 """ARGprism package."""
 
 from ._version import __version__
-from .pipeline import PipelineResult, run_pipeline
+from .core.pipeline import PipelineResult, run_pipeline
 
 __all__ = ["run_pipeline", "PipelineResult", "__version__"]
