+++ argprism/analysis/__init__.py
@@ -1,0 +1,6 @@
+"""Analysis and alignment components."""
+
+from .diamond_mapping import run_diamond
+from .reporter import generate_report, load_metadata, parse_diamond_hits
+
+__all__ = ["run_diamond", "generate_report", "load_metadata", "parse_diamond_hits"]
