+++ argprism/classifier/__init__.py
@@ -1,0 +1,12 @@
+"""Machine learning components for ARG prediction."""
+
+from .arg_classifier import ARGClassifier, classify_embeddings, load_classifier
+from .embeddings_generator import generate_embeddings, load_plm
+
+__all__ = [
+    "ARGClassifier",
+    "classify_embeddings", 
+    "load_classifier",
+    "generate_embeddings",
+    "load_plm"
+]
