+++ argprism/cli/__init__.py
@@ -1,0 +1,5 @@
+"""Command line interface components."""
+
+from .cli import main
+
+__all__ = ["main"]
